perf: inject social icons on demand
